pyproject.toml
[project]
name = "video-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.12.0",
    "beanie==2.0.1",
    "click==8.3.1",
    "dnspython==2.8.0",
    "fastapi==0.123.9",
    "graphql-core==3.3.0a11",
    "idna==3.11",
    "motor>=3.7.0",
    "pydantic==2.12.5",
    "pymongo==4.15.5",
    "starlette==0.50.0",
    "strawberry-graphql==0.287.2",
    "typing-extensions==4.15.0",
    "uvicorn==0.38.0",
    "pyyaml==6.0.3",
]
